Decoupled chess logic from game manager; Added game state controller; Added restart button when the game is over; Reworked the logic for square size; Reworked resource manager to avoid hard coding assets in the constructor; Fixed the promotion pieces having incorrect size

+++ Chess/Square.cpp
@@ -1,12 +1,11 @@
 #include "pch.h"
 #include "Square.h"
 #include "Colors.h"
-#include "GameManager.h"
 
 Square::Square(const sf::Color& color,
 	const sf::Vector2i& coordinates,
 	const sf::Vector2f& position,
-	double size,
+	float size,
 	const State& state
 	)
 {
